modify OneFragment.kt and some
