+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/CustomAdapter.kt
@@ -1,0 +1,41 @@
+package jp.co.yumemi.android.code_check
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+
+//OneFragmentで使用します
+class CustomAdapter(
+    private val itemClickListener: OnItemClickListener,
+) : ListAdapter<Item, CustomAdapter.ViewHolder>(diff_util){
+
+    class ViewHolder(view: View): RecyclerView.ViewHolder(view)
+
+    interface OnItemClickListener{
+        fun itemClick(item: Item)
+    }
+
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int,
+    ): ViewHolder {
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.layout_item, parent, false)
+        return ViewHolder(view)
+    }
+
+    override fun onBindViewHolder(
+        holder: ViewHolder,
+        position: Int,
+    ) {
+        val item = getItem(position)
+        val textView: TextView = holder.itemView.findViewById<TextView>(R.id.repositoryNameView)
+        textView.text = item.name
+
+        holder.itemView.setOnClickListener{
+            itemClickListener.itemClick(item)
+        }
+    }
+}
