+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/OneFragment.kt
@@ -10,7 +10,9 @@ import android.view.ViewGroup
 import android.view.inputmethod.EditorInfo
 import android.widget.TextView
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.LifecycleCoroutineScope
 import androidx.lifecycle.lifecycleScope
+import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.*
 import jp.co.yumemi.android.code_check.databinding.FragmentOneBinding
@@ -29,97 +31,79 @@ class OneFragment: Fragment(R.layout.fragment_one){
         val viewModel = OneViewModel(context)
         val layoutManager = LinearLayoutManager(context)
         val dividerItemDecoration = DividerItemDecoration(context, layoutManager.orientation)
+
         val adapter= CustomAdapter(
             object : CustomAdapter.OnItemClickListener{
                 override fun itemClick(item: Item){
-                    gotoRepositoryFragment(item)
+                    goToRepositoryFragment(item, findNavController())
                 }
             }
         )
 
-        binding.searchInputText.setOnEditorActionListener { editText, action, _ ->
-            if (action == EditorInfo.IME_ACTION_SEARCH) {
-                val query = editText.text.toString()
-                performSearch(query, viewModel, adapter)
-                true
-            } else {
-                false
-            }
-        }
+        bindingHandler(binding, viewModel, adapter, lifecycleScope, layoutManager, dividerItemDecoration)
 
-        binding.recyclerView.also{
-            it.layoutManager = layoutManager
-            it.addItemDecoration(dividerItemDecoration)
-            it.adapter = adapter
-        }
     }
 
-    fun gotoRepositoryFragment(
-        item: Item
-    ) {
-        val action= OneFragmentDirections.actionRepositoriesFragmentToRepositoryFragment(item)
-        findNavController().navigate(action)
-    }
+}
 
+private fun goToRepositoryFragment(
+    item: Item,
+    findNavController: NavController,
+) {
+    val action= OneFragmentDirections.actionRepositoriesFragmentToRepositoryFragment(item)
+    findNavController.navigate(action)
+}
 
-    private fun performSearch(
-        query: String,
-        viewModel: OneViewModel,
-        adapter: CustomAdapter
-    ) {
-        lifecycleScope.launch {
-            val searchResult = viewModel.searchResults(query)
-            adapter.submitList(searchResult)
+private fun bindingHandler(
+    binding: FragmentOneBinding,
+    viewModel: OneViewModel,
+    adapter: CustomAdapter,
+    lifecycleScope: LifecycleCoroutineScope,
+    layoutManager: RecyclerView.LayoutManager,
+    dividerItemDecoration: DividerItemDecoration,
+){
+    binding.searchInputText.setOnEditorActionListener { editText, action, _ ->
+        if (action == EditorInfo.IME_ACTION_SEARCH) {
+            val query = editText.text.toString()
+            performSearch(query, viewModel, adapter, lifecycleScope)
+            true
+        } else {
+            false
         }
     }
 
+    binding.recyclerView.also{
+        it.layoutManager = layoutManager
+        it.addItemDecoration(dividerItemDecoration)
+        it.adapter = adapter
+    }
+}
+
+private fun performSearch(
+    query: String,
+    viewModel: OneViewModel,
+    adapter: CustomAdapter,
+    lifecycleScope: LifecycleCoroutineScope,
+) {
+    lifecycleScope.launch {
+        val searchResult = viewModel.searchResults(query)
+        adapter.submitList(searchResult)
+    }
 }
 
 val diff_util = object: DiffUtil.ItemCallback<Item>(){
     override fun areItemsTheSame(
         oldItem: Item,
-        newItem: Item
+        newItem: Item,
     ): Boolean {
         return oldItem.name == newItem.name
     }
 
     override fun areContentsTheSame(
         oldItem: Item,
-        newItem: Item
+        newItem: Item,
     ): Boolean {
         return oldItem == newItem
     }
 
-}
-
-class CustomAdapter(
-    private val itemClickListener: OnItemClickListener,
-) : ListAdapter<Item, CustomAdapter.ViewHolder>(diff_util){
-
-    class ViewHolder(view: View): RecyclerView.ViewHolder(view)
-
-    interface OnItemClickListener{
-    	fun itemClick(item: Item)
-    }
-
-    override fun onCreateViewHolder(
-        parent: ViewGroup,
-        viewType: Int,
-    ): ViewHolder {
-    	val view = LayoutInflater.from(parent.context).inflate(R.layout.layout_item, parent, false)
-    	return ViewHolder(view)
-    }
-
-    override fun onBindViewHolder(
-        holder: ViewHolder,
-        position: Int,
-    ) {
-    	val item = getItem(position)
-        val textView: TextView = holder.itemView.findViewById<TextView>(R.id.repositoryNameView)
-        textView.text = item.name
-
-    	holder.itemView.setOnClickListener{
-     		itemClickListener.itemClick(item)
-    	}
-    }
 }
