+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/OneViewModel.kt
@@ -74,13 +74,3 @@ class OneViewModel(
     }
 }
 
-@Parcelize
-data class Item(
-    val name: String,
-    val ownerIconUrl: String,
-    val language: String,
-    val stargazersCount: Long,
-    val watchersCount: Long,
-    val forksCount: Long,
-    val openIssuesCount: Long,
-) : Parcelable
