+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/TopActivity.kt
@@ -3,11 +3,7 @@
  */
 package jp.co.yumemi.android.code_check
 
-import android.os.Build
-import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
-import java.time.LocalDate
-import java.time.ZoneId
 import java.util.*
 
 class TopActivity : AppCompatActivity(R.layout.activity_top) {
